Add sample producer (not for lattice QCD lib, but for an experiment) Add Make writer (trying to cross platform, far from done)

+++ Code/CLGLib/Measurement/GaugeSmearing/CGaugeSmearingAPEProj.cu
@@ -68,7 +68,7 @@ void CGaugeSmearingAPEProj::Initial(class CLatticeData* pOwner, const CParameter
     }
     else if (iValue >= 4 && iValue < 255)
     {
-        m_byProjIterate = static_cast<UINT>(iValue);
+        m_byProjIterate = static_cast<BYTE>(iValue);
     }
     else
     {
